The Rust fn-name spelling moves to almide-base so both test legs can read one predicate
